backend/requirements.txt
fastapi
email-validator
uvicorn[standard]
python-multipart
google-genai
google-auth-oauthlib
google-api-python-client
python-dotenv
pytesseract
pillow
pypdf
pypdfium2
httpx
mammoth
python-pptx
recost

# OCR pipeline (Docling = default, GOT-OCR 2.0 = optional GPU fallback)
docling>=2.15,<3
transformers>=4.46,<5
torch>=2.5,<3  # install CPU wheel in CPU-only deployments: pip install torch --index-url https://download.pytorch.org/whl/cpu

openpyxl
beautifulsoup4
python-Levenshtein

# Testing
pytest
pytest-anyio
email-validator

# Pydantic AI agents (see docs/decisions/0001-adopt-pydantic-ai.md)
pydantic-ai-slim[google]>=0.0.20
logfire>=2.0
sse-starlette>=2.0

# Optional (evals only — not loaded at runtime). Install with:
#   pip install 'pydantic-evals>=0.0.5'
# Required by backend/tests/evals/document_classification.py.
# pydantic-evals>=0.0.5
